+++ data_access/src/main/java/edu/duke/ece651/team1/data_access/Student/StudentDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface StudentDao {
-    void addStudent(Student student, long userId);
-    void removeStudent(long studentID);
-    Optional<Student> findStudentByStudentID(long studentID);
-    Optional<Student> findStudentByUserID(long userID);
+    void addStudent(Student student, int userId);
+    void removeStudent(int studentID);
+    Optional<Student> findStudentByStudentID(int studentID);
+    Optional<Student> findStudentByUserID(int userID);
     List<Student> getAllStudents();
 }
