+++ server/src/main/java/edu/duke/ece651/team1/server/service/UserService.java
@@ -67,7 +67,7 @@ public class UserService implements UserDetailsService {
     public void createUserStudent(String username, String password,String role,String legalName, String displayName,String email) {
         User user = new User(username,passwordEncoder.encode(password),role);
         userDao.addUser(user);
-        long userId = userDao.findUserByUsername(username).getUserId();
+        int userId = userDao.findUserByUsername(username).getUserId();
         Student student = new Student(legalName,displayName,email);
         studentDao.addStudent(student,userId);
